// src/modules/edc-dmgmt-client/model/permission.ts
/**
 * EDC REST API
 * All files merged by open api merger
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Action } from './action';
import { Constraint } from './constraint';
import { Duty } from './duty';
import {Rule} from "./rule";


export interface Permission extends Rule {
    duties?: Array<Duty>;
}